+++ spectral_cube/cube_utils.py
@@ -256,8 +256,11 @@ def try_load_beam(header):
         beam = Beam.from_fits_header(header)
         return beam
     except Exception as ex:
-        warnings.warn("Could not parse beam information from header."
-                      "  Exception was: {0}".format(ex.__repr__()))
+        # We don't emit a warning if no beam was found since it's ok for
+        # cubes to not have beams
+        if 'No BMAJ' not in str(ex):
+            warnings.warn("Could not parse beam information from header."
+                          "  Exception was: {0}".format(ex.__repr__()))
 
 
 def beams_to_bintable(beams):